+++ mobile/android/android-components/components/support/ktx/src/main/java/mozilla/components/support/ktx/android/content/Context.kt
@@ -11,6 +11,7 @@ import android.content.Intent
 import android.content.Intent.ACTION_SEND
 import android.content.Intent.EXTRA_SUBJECT
 import android.content.Intent.EXTRA_TEXT
+import android.content.Intent.EXTRA_TITLE
 import android.content.Intent.FLAG_ACTIVITY_NEW_TASK
 import android.content.pm.PackageManager.PERMISSION_GRANTED
 import android.hardware.camera2.CameraManager
@@ -71,9 +72,9 @@ fun Context.hasCamera(): Boolean {
 }
 
 /**
- *  Shares content via [ACTION_SEND] intent.
+ * Shares content via [ACTION_SEND] intent.
  *
- * @param text the data to be shared  [EXTRA_TEXT]
+ * @param text the data to be shared [EXTRA_TEXT]
  * @param subject of the intent [EXTRA_TEXT]
  * @return true it is able to share false otherwise.
  */
@@ -81,6 +82,7 @@ fun Context.share(text: String, subject: String = getString(R.string.mozac_suppo
     return try {
         val intent = Intent(ACTION_SEND).apply {
             type = "text/plain"
+            putExtra(EXTRA_TITLE, subject)
             putExtra(EXTRA_SUBJECT, subject)
             putExtra(EXTRA_TEXT, text)
             flags = FLAG_ACTIVITY_NEW_TASK